backend/requirements.txt
flask
flask-cors
openai
requests
together
pandas
beautifulsoup4
PyMuPDF
openpyxl
python-pptx
python-dotenv
mcp
httpx 
gen3
pydicom
Pillow
numpy
requests
python-gdcm
pylibjpeg
pylibjpeg-libjpeg
pylibjpeg-openjpeg